random number generator and f64 to u64

+++ src/bubble_sort.rs
@@ -1,7 +1,7 @@
-pub fn serial_bubble_sort(arr: &mut [f64]){
+pub fn serial_bubble_sort(arr: &mut [u64]) {
     let n: usize = arr.len();
-    for i in 0..n-1 {
-        for j in 0..n-i-1 {
+    for i in 0..n - 1 {
+        for j in 0..n - i - 1 {
             if arr[j] > arr[j + 1] {
                 let temp = arr[j];
                 arr[j] = arr[j + 1];
@@ -11,6 +11,6 @@ pub fn serial_bubble_sort(arr: &mut [f64]){
     }
 }
 
-// pub fn parallel_bubble_sort(arr: &mut [f64]) -> &mut [f64]{
+// pub fn parallel_bubble_sort(arr: &mut [u64]) -> &mut [u64]{
 //     return arr;
 // }
